requirements.txt
pip==23.3.1
werkzeug==2.3.7
flask==2.3.3
flask-socketio==5.3.6
numpy>=1.26.0
openpyxl==3.1.2
python-dotenv==1.0.0
boto3==1.28.44
python-engineio==4.8.0
python-socketio==5.10.0
tqdm==4.65.0
polars==0.18.0
# 添加一些必要的间接依赖
blinker>=1.6.2
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
python-dateutil>=2.8.2
pytz>=2020.1
six>=1.5 
xlsx2csv==0.8.4